pyproject.toml
[tool.ruff]
fix = true
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

line-length = 80

[tool.ruff.lint]
preview = true
select = ["E", "F", "B"]
#select = ["E", "F", "I", "PL", "I", "B"]
ignore = ["B008", "E501"]  # black will take care of line-length violations
ignore-init-module-imports = true
unfixable = []

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403", "F405", "E501"]
"consts.py" = ["E501"]

[tool.black]
line-length = 80
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
[tool.isort]
profile = "black"
line_length = 80

#multi_line_output = 3
#include_trailing_comma = true
#force_grid_wrap = 0

[project]
name = "proxy_getter"
authors = [
    { name = "Francis Secada", email = "francis.secada@gmail.com" }
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
version = "0.0.1b1"

dependencies = [
    "free-proxy>=1.1.3",
    "geopy>=2.4.1",
    "html5lib>=1.1",
    "httpx[socks]>=0.28.1",
    "loguru>=0.7.3",
    "playwright>=1.52.0",
    "sqlalchemy-mixins>=2.0.5",
    "sqlmodel>=0.0.24",
    "sqlmodel-crud-utilities>=0.1.0",
]

[dependency-groups]
dev = [
    "alembic>=1.15.2",
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["proxy_getter"]
name_spaces = false
